base: Move the deserialized responses types into the common crate
