+++ matrix_sdk_base/Cargo.toml
@@ -25,7 +25,7 @@ docs = ["encryption", "sled_cryptostore", "messages"]
 
 [dependencies]
 dashmap= "4.0.1"
-serde = { version = "1.0.118", features = ["rc"]}
+serde = { version = "1.0.119", features = ["rc"] }
 serde_json = "1.0.61"
 zeroize = "1.2.0"
 tracing = "0.1.22"
