+++ matrix_sdk_common/src/lib.rs
@@ -13,6 +13,7 @@ pub use ruma::{
 
 pub use uuid;
 
+pub mod deserialized_responses;
 pub mod locks;
 
 /// Super trait that is used for our store traits, this trait will differ if
