+++ matrix_sdk_crypto/Cargo.toml
@@ -25,7 +25,7 @@ matrix-sdk-common = { version = "0.2.0", path = "../matrix_sdk_common" }
 
 olm-rs = { version = "1.0.0", features = ["serde"] }
 getrandom = "0.2.1"
-serde = { version = "1.0.118", features = ["derive", "rc"] }
+serde = { version = "1.0.119", features = ["derive", "rc"] }
 serde_json = "1.0.61"
 zeroize = { version = "1.2.0", features = ["zeroize_derive"] }
