+++ matrix_sdk_test/Cargo.toml
@@ -16,4 +16,4 @@ http = "0.2.2"
 matrix-sdk-common = { version = "0.2.0", path = "../matrix_sdk_common" }
 matrix-sdk-test-macros = { version = "0.1.0", path = "../matrix_sdk_test_macros" }
 lazy_static = "1.4.0"
-serde = "1.0.118"
+serde = "1.0.119"
